Add accounts editing and remove auto-animate lib

+++ leetnode/src/pages/api/user/admin/deleteMany.ts
@@ -1,18 +1,37 @@
 import { NextApiRequest, NextApiResponse } from "next";
+import { getServerSession } from "next-auth/next";
 
+import { authOptions } from "@/pages/api/auth/[...nextauth]";
 import { prisma } from "@/server/db/client";
+import { Role } from "@prisma/client";
 
 export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse
 ) {
+  const session = await getServerSession(req, res, authOptions);
   const emails: string[] = req.body.emails;
+  if (emails.includes(session?.user?.email as string)) {
+    return res.status(403).json({
+      message: `Not allowed to delete yourself`,
+    });
+  }
 
   try {
     await prisma.user.deleteMany({
       where: {
         email: {
           in: emails,
+        },
+        AND: {
+          role: {
+            not: Role.SUPERUSER,
+          },
+          OR: {
+            role: {
+              not: Role.ADMIN,
+            },
+          },
         },
       },
     });
@@ -22,6 +41,6 @@ export default async function handler(
   }
 
   res.status(200).json({
-    message: "Users deleted successfully",
+    message: "Users deleted successfully (admins ignored)",
   });
 }
